fixed anthro tables

+++ packages/anthro/anthro/utils/reach_vegetation.py
@@ -75,7 +75,7 @@ def vegetation_summary(outputs_gpkg_path: str, dgo: str, veg_raster: str):
                 for oldvalue in np.unique(mask_raster):
                     if oldvalue is not np.ma.masked:
                         cell_count = np.count_nonzero(mask_raster == oldvalue)
-                        veg_counts.append([reach_id, int(oldvalue), cell_count * cell_area, cell_count])
+                        veg_counts.append([reach_id, int(oldvalue), float(cell_count * cell_area), int(cell_count)])
             except Exception as ex:
                 log.warning('Error obtaining vegetation raster values for ReachID {}'.format(reach_id))
                 log.warning(ex)
